requirements.txt
flask==3.0.3
python-dotenv==1.0.1
openai==1.43.0
pydantic==2.8.2
httpx==0.27.2
